+++ scripts/module.js
@@ -23,6 +23,7 @@ import {settingButton} from './settingsMenu.js';
 import {remoteDialog, remoteDocumentDialog, remoteDocumentsDialog} from './utility/remoteDialog.js';
 import {diceSoNice} from './diceSoNice.js';
 import {info, setCompendiumItemInfo, stripUnusedFlags} from './info.js';
+import {applyEquipmentFlag, itemFeatures, itemFeaturesDelete} from './equipment.js';
 export let socket;
 Hooks.once('init', async function() {
     registerSettings();
@@ -192,10 +193,15 @@ Hooks.once('ready', async function() {
     }
     if (game.settings.get('chris-premades', 'Arcane Ward')) Hooks.on('midi-qol.damageApplied', macros.arcaneWard.damage);
     if (game.settings.get('chris-premades', 'Automation Verification')) Hooks.on('midi-qol.preItemRoll', info);
+    if (game.settings.get('chris-premades', 'Item Features')) {
+        Hooks.on('preUpdateItem', itemFeatures);
+        Hooks.on('preDeleteItem', itemFeaturesDelete);
+    }
 });
 let dev = {
     'setCompendiumItemInfo': setCompendiumItemInfo,
-    'stripUnusedFlags': stripUnusedFlags
+    'stripUnusedFlags': stripUnusedFlags,
+    'applyEquipmentFlag': applyEquipmentFlag
 }
 globalThis['chrisPremades'] = {
     helpers,